+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - feat: Set `max_line_length` in `.editorconfig`
+- fix: Make Prettier follow `.editorconfig`
 
 ## 3.0.0 – 3.0.1